Integrated extra tests for 68k and fixed some of them but it still fails 80% =(
Tests from https://github.com/SingleStepTests/m68000

+++ src/instruction_decoder.cpp
@@ -4,6 +4,46 @@
 #include <functional>
 #include <memory>
 #include <cstdint>
+#include "instructions/move.hpp"
+#include "instructions/moveq.hpp"
+#include "instructions/lea.hpp"
+#include "instructions/pea.hpp"
+#include "instructions/add.hpp"
+#include "instructions/adda.hpp"
+#include "instructions/addx.hpp"
+#include "instructions/addi.hpp"
+#include "instructions/addq.hpp"
+#include "instructions/sub.hpp"
+#include "instructions/suba.hpp"
+#include "instructions/subx.hpp"
+#include "instructions/subi.hpp"
+#include "instructions/subq.hpp"
+#include "instructions/and.hpp"
+#include "instructions/andi.hpp"
+#include "instructions/or.hpp"
+#include "instructions/ori.hpp"
+#include "instructions/eor.hpp"
+#include "instructions/eori.hpp"
+#include "instructions/clr.hpp"
+#include "instructions/neg.hpp"
+#include "instructions/bit_manip.hpp"
+#include "instructions/bit_shift.hpp"
+#include "instructions/mul.hpp"
+#include "instructions/div.hpp"
+#include "instructions/illegal.hpp"
+#include "instructions/nop.hpp"
+#include "instructions/jmp.hpp"
+#include "instructions/jsr.hpp"
+#include "instructions/rts.hpp"
+#include "instructions/link.hpp"
+#include "instructions/unlk.hpp"
+#include "instructions/ext.hpp"
+#include "instructions/bcc.hpp"
+#include "instructions/scc.hpp"
+#include "instructions/tst.hpp"
+#include "instructions/cmp.hpp"
+#include "instructions/cmpa.hpp"
+#include "instructions/cmpi.hpp"
 
 using namespace M68K;
 
@@ -123,7 +163,7 @@ void InstructionDecoder::generateOpcodeTable(){
             break;
         }
         if (!instruction)
-            instruction = INSTRUCTION::Illegal::create(opcode);
+            instruction = INSTRUCTION::Illegal::create((uint16_t)opcode);
         this->opcode_table.push_back(std::move(instruction));
     }
 }
